requirements.txt
# file: requirements.txt
rich>=13.0.0
pyyaml>=6.0
requests>=2.31.0
pefile>=2023.2.7
yara-python>=4.3.1
python-magic>=0.4.27